+++ src/controller/IAppControllerVToM.java
@@ -3,8 +3,8 @@ package controller;
 import java.util.List;
 import java.util.Optional;
 
-import kide.KideAppEvent;
-import model.KideAppEventDetails;
+import model.kide.KideAppEvent;
+import model.kide.KideAppEventDetails;
 
 public interface IAppControllerVToM {
 	void receiveEvents(Optional<List<KideAppEvent>> events);
