+++ src/view/Gui.java
@@ -13,8 +13,8 @@ import javafx.scene.layout.AnchorPane;
 import javafx.scene.layout.BorderPane;
 import javafx.scene.layout.HBox;
 import javafx.stage.Stage;
-import kide.KideAppEvent;
-import model.KideAppEventDetails;
+import model.kide.KideAppEvent;
+import model.kide.KideAppEventDetails;
 import view.layout.base.BaseLayoutController;
 import view.layout.notification.NotificationLayoutController;
 import view.layout.upsert.UpsertLayoutController;
