+++ src/view/IGui.java
@@ -4,8 +4,8 @@ import java.util.List;
 import java.util.Optional;
 
 import controller.IAppControllerMToV;
-import kide.KideAppEvent;
-import model.KideAppEventDetails;
+import model.kide.KideAppEvent;
+import model.kide.KideAppEventDetails;
 
 public interface IGui {
 	void handleEvents(Optional<List<KideAppEvent>> events);
