+++ src/view/layout/statistics/datastrategy/IDateGetter.java
@@ -2,7 +2,7 @@ package view.layout.statistics.datastrategy;
 
 import java.time.ZonedDateTime;
 
-import kide.KideAppEvent;
+import model.kide.KideAppEvent;
 
 public interface IDateGetter {
 	public ZonedDateTime get(KideAppEvent e);
