+++ test/kide/KideAppApiTest.java
@@ -10,6 +10,9 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import model.kide.KideAppApi;
+import model.kide.KideAppEvent;
+
 class KideAppApiTest {
 	KideAppApi api;
